Create hourly-project-price-setter
 The setting below prevents an update of the previous price setting. If the rate of a worker is changed, the settings make sure that the new rate does not the previous rates. For example, if a worker was working on a rate of $20/hr on Mon, Tue, Wed, and then promoted on Thurs to work at $30/hr. The rate of Mon, Tue, and Wed remains at $20. 
The if statement, ensures shifts that don't meet the condition of greater than zero are assigned the set rate for the worker. We do this by going to the shiftpricesetter table. We then grab the worker, grab the project, business, and a record where the setting for using is set to 'Yes Using' or the current price.  We cannot set one price for worker because different projects can have different prices